encapsulate labels
